fix: convert uint32 to uint256 when calculate fraction

+++ contracts/ManagementFund.sol
@@ -15,10 +15,7 @@ contract ManagementFund {
         Governance g = Governance(governanceContract);
         ERC20 t = ERC20(g.getARATokenContract());
 
-        require(
-            t.balanceOf(address(this)) > 0,
-            "60"
-        );
+        require(t.balanceOf(address(this)) > 0, "60");
 
         uint256 managementFund = t.balanceOf(address(this));
         for (uint16 i = 0; i < g.getTotalManager(); i++) {
@@ -27,8 +24,8 @@ contract ManagementFund {
                 g.getManager(i).controlWeight > 0
             ) {
                 uint256 amount = (managementFund *
-                    g.getManager(i).controlWeight) /
-                    g.getManagerMaxControlWeight();
+                    uint256(g.getManager(i).controlWeight)) /
+                    uint256(g.getManagerMaxControlWeight());
                 t.transferFrom(address(this), g.getManager(i).addr, amount);
             }
         }
